+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Moved wayland tracking to separate crate and published as `gayland`.
 - Update gui size for proper dialog scaling.
 - Add pages to gui for configuring capture and recording control.
+- Add runtime libinput based hotkey support.
 
 ## [0.1.2] - 2026-04-27
 ### Changed